+++ src/components/formulaire/Formulaire.tsx
@@ -29,21 +29,25 @@ export function Formulaire(props: OrchestratedElement) {
 	);
 
 	return (
-		<form>
-			{components.map(function (component: ComponentType) {
-				const { componentType, id } = component;
+		<div className="fr-grid-row fr-grid-row--center fr-grid-row--middle fr-mt-6w fr-mt-md-12w">
+			<div className="fr-col-lg-6 fr-col-12">
+				<form>
+					{components.map(function (component: ComponentType) {
+						const { componentType, id } = component;
 
-				if (componentType in lunatic) {
-					const Component = lunatic[componentType];
+						if (componentType in lunatic) {
+							const Component = lunatic[componentType];
 
-					return (
-						<LunaticComponentContainer key={id} id={id}>
-							<Component key={id} {...component} errors={errors} />
-						</LunaticComponentContainer>
-					);
-				}
-				return null;
-			})}
-		</form>
+							return (
+								<LunaticComponentContainer key={id} id={id}>
+									<Component key={id} {...component} errors={errors} />
+								</LunaticComponentContainer>
+							);
+						}
+						return null;
+					})}
+				</form>
+			</div>
+		</div>
 	);
 }
