--- src/components/navigation/index.ts
@@ -1,1 +1,0 @@
-export { Navigation } from './Navigation';
